Separate ST logic and remove all constructors and builders with startEngine
